sandbox/pyproject.toml
[project]
name = "gestalt-sandbox"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "anthropic>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
